# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/pools": {
      "post": {
        "operationId": "PoolController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoolDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pool created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pool"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Pool already exists at this address/chain"
          }
        },
        "summary": "Create a new Pool",
        "tags": [
          "pools"
        ]
      },
      "get": {
        "operationId": "PoolController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all Pools",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pool"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all Pools with their relations",
        "tags": [
          "pools"
        ]
      }
    },
    "/pools/cross-chain": {
      "get": {
        "description": "Given a token address on a specific chain, finds all pools (on any chain) that contain this canonical token",
        "operationId": "PoolController_findCrossChainPools",
        "parameters": [
          {
            "name": "tokenAddress",
            "required": true,
            "in": "query",
            "description": "Token address",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chainId",
            "required": true,
            "in": "query",
            "description": "Chain ID where token exists",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pools across all chains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pool"
                  }
                }
              }
            }
          }
        },
        "summary": "Find all pools containing a token across all chains",
        "tags": [
          "pools"
        ]
      }
    },
    "/pools/on-chain": {
      "get": {
        "operationId": "PoolController_findPoolsOnChain",
        "parameters": [
          {
            "name": "tokenAddress",
            "required": true,
            "in": "query",
            "description": "Token address",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chainId",
            "required": true,
            "in": "query",
            "description": "Chain ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pools on the specified chain",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pool"
                  }
                }
              }
            }
          }
        },
        "summary": "Find pools on a specific chain containing a token",
        "tags": [
          "pools"
        ]
      }
    },
    "/pools/by-pair": {
      "get": {
        "operationId": "PoolController_findPoolsByPair",
        "parameters": [
          {
            "name": "token0Address",
            "required": true,
            "in": "query",
            "description": "First token address",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token1Address",
            "required": true,
            "in": "query",
            "description": "Second token address",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chainId",
            "required": true,
            "in": "query",
            "description": "Chain ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pools for this token pair",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pool"
                  }
                }
              }
            }
          }
        },
        "summary": "Find all pools for a token pair on a specific chain",
        "tags": [
          "pools"
        ]
      }
    },
    "/pools/by-address": {
      "get": {
        "operationId": "PoolController_findPoolByAddress",
        "parameters": [
          {
            "name": "poolAddress",
            "required": true,
            "in": "query",
            "description": "Pool contract address",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chainId",
            "required": true,
            "in": "query",
            "description": "Chain ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pool found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pool"
                }
              }
            }
          },
          "404": {
            "description": "Pool not found"
          }
        },
        "summary": "Find pool by its address on a specific chain",
        "tags": [
          "pools"
        ]
      }
    },
    "/pools/{id}": {
      "get": {
        "operationId": "PoolController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Pool UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pool found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pool"
                }
              }
            }
          },
          "404": {
            "description": "Pool not found"
          }
        },
        "summary": "Get a Pool by ID with its relations",
        "tags": [
          "pools"
        ]
      },
      "put": {
        "operationId": "PoolController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Pool UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePoolDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pool updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pool"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Pool not found"
          },
          "409": {
            "description": "Pool already exists at this address/chain"
          }
        },
        "summary": "Update a Pool",
        "tags": [
          "pools"
        ]
      },
      "delete": {
        "operationId": "PoolController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Pool UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Pool deleted successfully"
          },
          "404": {
            "description": "Pool not found"
          }
        },
        "summary": "Delete a Pool",
        "tags": [
          "pools"
        ]
      }
    },
    "/dexes": {
      "post": {
        "operationId": "DexController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDexDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "DEX created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dex"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "DEX with same name already exists"
          }
        },
        "summary": "Create a new DEX",
        "tags": [
          "dexes"
        ]
      },
      "get": {
        "operationId": "DexController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all DEXes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dex"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all DEXes",
        "tags": [
          "dexes"
        ]
      }
    },
    "/dexes/{id}": {
      "get": {
        "operationId": "DexController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "DEX UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "DEX found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dex"
                }
              }
            }
          },
          "404": {
            "description": "DEX not found"
          }
        },
        "summary": "Get a DEX by ID",
        "tags": [
          "dexes"
        ]
      },
      "put": {
        "operationId": "DexController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "DEX UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDexDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DEX updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dex"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "DEX not found"
          },
          "409": {
            "description": "DEX with same name already exists"
          }
        },
        "summary": "Update a DEX",
        "tags": [
          "dexes"
        ]
      },
      "delete": {
        "operationId": "DexController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "DEX UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "DEX deleted successfully"
          },
          "404": {
            "description": "DEX not found"
          }
        },
        "summary": "Delete a DEX",
        "tags": [
          "dexes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Arbi Engine API",
    "description": "API documentation for Arbi Engine operations",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreatePoolDto": {
        "type": "object",
        "properties": {
          "poolAddress": {
            "type": "string",
            "description": "Pool contract address"
          },
          "chainId": {
            "type": "number",
            "description": "Blockchain chain ID"
          },
          "token0Id": {
            "type": "string",
            "description": "Token 0 canonical ID (UUID)"
          },
          "token0Address": {
            "type": "string",
            "description": "Token 0 address on the chain"
          },
          "token1Id": {
            "type": "string",
            "description": "Token 1 canonical ID (UUID)"
          },
          "token1Address": {
            "type": "string",
            "description": "Token 1 address on the chain"
          },
          "dexId": {
            "type": "string",
            "description": "DEX ID (UUID)"
          },
          "fee": {
            "type": "number",
            "description": "Pool fee in basis points"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata as JSON object"
          }
        },
        "required": [
          "poolAddress",
          "chainId",
          "token0Id",
          "token0Address",
          "token1Id",
          "token1Address",
          "dexId"
        ]
      },
      "Pool": {
        "type": "object",
        "properties": {}
      },
      "UpdatePoolDto": {
        "type": "object",
        "properties": {
          "poolAddress": {
            "type": "string",
            "description": "Pool contract address"
          },
          "chainId": {
            "type": "number",
            "description": "Blockchain chain ID"
          },
          "token0Id": {
            "type": "string",
            "description": "Token 0 canonical ID (UUID)"
          },
          "token0Address": {
            "type": "string",
            "description": "Token 0 address on the chain"
          },
          "token1Id": {
            "type": "string",
            "description": "Token 1 canonical ID (UUID)"
          },
          "token1Address": {
            "type": "string",
            "description": "Token 1 address on the chain"
          },
          "dexId": {
            "type": "string",
            "description": "DEX ID (UUID)"
          },
          "fee": {
            "type": "number",
            "description": "Pool fee in basis points"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata as JSON object"
          }
        }
      },
      "CreateDexDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique name of the DEX"
          },
          "swapTopic": {
            "type": "string",
            "description": "Event topic hash for swap events"
          },
          "addLiquidityTopic": {
            "type": "string",
            "description": "Event topic hash for add liquidity events"
          },
          "removeLiquidityTopic": {
            "type": "string",
            "description": "Event topic hash for remove liquidity events"
          }
        },
        "required": [
          "name",
          "swapTopic"
        ]
      },
      "Dex": {
        "type": "object",
        "properties": {}
      },
      "UpdateDexDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique name of the DEX"
          },
          "swapTopic": {
            "type": "string",
            "description": "Event topic hash for swap events"
          },
          "addLiquidityTopic": {
            "type": "string",
            "description": "Event topic hash for add liquidity events"
          },
          "removeLiquidityTopic": {
            "type": "string",
            "description": "Event topic hash for remove liquidity events"
          }
        }
      },
...
    }
  }
}